Job Description README.md
Rocket Financial is an early-stage start-up in the Financial Technology space. Our mission is to change the way money moves around the planet. Primarily, we are building a business technology platform for those wanting to integrate and embed financial services into their customer applications.


We are looking for a Senior DevOps Engineer to join our team. You will be a core member working with the Product, and Engineering teams to deliver and maintain our environments and platform